Switch to float32

+++ individual_experiments.py
@@ -89,12 +89,12 @@ if __name__ == '__main__':
 
         X_shape = X_train.shape[1]
     elif os.path.exists(f'{args.experiment_dir}/{args.feature}/X_train'):
-        X_train = np.load(f'{args.experiment_dir}/{args.feature}/X_train.npy')
-        y_train = np.load(f'{args.experiment_dir}/{args.feature}/y_train.npy')
-        X_val = np.load(f'{args.experiment_dir}/{args.feature}/X_val.npy')
-        y_val = np.load(f'{args.experiment_dir}/{args.feature}/y_val.npy')
-        X_test = np.load(f'{args.experiment_dir}/{args.feature}/X_test.npy')
-        y_test = np.load(f'{args.experiment_dir}/{args.feature}/y_test.npy')
+        X_train = np.load(f'{args.experiment_dir}/{args.feature}/X_train.npy').astype(np.float32)
+        y_train = np.load(f'{args.experiment_dir}/{args.feature}/y_train.npy').astype(np.float32)
+        X_val = np.load(f'{args.experiment_dir}/{args.feature}/X_val.npy').astype(np.float32)
+        y_val = np.load(f'{args.experiment_dir}/{args.feature}/y_val.npy').astype(np.float32)
+        X_test = np.load(f'{args.experiment_dir}/{args.feature}/X_test.npy').astype(np.float32)
+        y_test = np.load(f'{args.experiment_dir}/{args.feature}/y_test.npy').astype(np.float32)
 
         X_shape = X_train.shape[1]
     else:
